add pagelayout component, apply to buy/swap
